// src/app/models/vaisseau.ts
export class Vaisseau {
    id:number;
    MGLT: string;
    cargo_capacity: string;
    consumables: string;
    cost_in_credits: string;
    created: string;
    crew: string;
    edited: string;
    hyperdrive_rating: string;
    length: string;
    manufacturer: string;
    max_atmosphering_speed: number;
    model: string;
    name: string;
    passengers: string;
    films: [];
    pilots: [];
    starship_class: string;
    url: string;
}